refactored authentication functions into new file
