+++ services/search.js
@@ -1,56 +1,7 @@
-const client_id = process.env.CLIENT_ID;
-const client_secret = process.env.CLIENT_SECRET;
 const axios = require('axios');
-const SearchParams = require('url').URLSearchParams;;
-const { ClientCredentials } = require('simple-oauth2')
+const SearchParams = require('url').URLSearchParams;
 const LOG = require('../utils/logger')
-
-const auth_config = {
-  client: {
-    id: client_id,
-    secret:  client_secret,
-  },
-  auth: {
-    tokenHost: 'https://accounts.spotify.com',
-          tokenPath: '/api/token'
-  }
-}
-
-const client = new ClientCredentials(auth_config); // we are using client credentials authorisation
-
-let access_token = null;
-
-// converts access token to the authorization header => { 'authorization': 'Bearer jsglksgkjlflsdkfssfkffs'}
-const TokenHeader = (access_token) => ({ 'Authorization': `${access_token.token.token_type} ${access_token.token.access_token}` });
-
-async function NewToken()
-{ // Return valid access token. called when token is expired or on first request
-  try {
-    LOG('getting New token');
-    const access_token = await client.getToken();
-    LOG('done..');
-
-    return access_token
-  }
-  catch (error) {
-    console.log("Error getting new token::>", error.output);
-  }
-}
-
-async function NewGetToken()
-{ // should return valid token
-  // everything being used is the access token which is an object containing the actual token and other things
-  LOG('getting valid token');
-
-  if ( !access_token || access_token.expired()) { // if the access token is null - on first request or when the token has expired. this methods eliminates the need for me to check that manually..
-    access_token = await NewToken()
-  }
-  
-  console.log('token::>', access_token.token.access_token);
-  console.log('token expiry time::>', access_token.token.expires_at);
-
-  return access_token
-}
+const { GetTokenHeader } = require('./auth')
 
 
 function GetRelevantKeys(body)
@@ -93,20 +44,18 @@ async function search(body)
   const alt_quer = new SearchParams(params);
   const req_query = alt_quer.toString();
 
-  console.log('request query::', req_query);
+  console.log('Request query::>', req_query);
 
-  var url = 'https://api.spotify.com/v1/search?' + req_query;
-
-  const access_token = await NewGetToken(); // returns access token as opposed to the actuall token ; access token contains the actual token and other useful shit
-  
-  var options = {
-    headers: TokenHeader(access_token),
+  const url = 'https://api.spotify.com/v1/search?' + req_query;
+  const token_header = await GetTokenHeader()
+  const options = {
+    headers: token_header,
     json: true
   };
 
-  var result = await axios.get(url, options);
+  const result = await axios.get(url, options);
+  console.log('Spotify response::>', result.status);
 
-  console.log('spotify response::\n', result.data);
   LOG('done quering spotify');
   return result.data;
 }
